Early implementation of SPI LED CoProcessor
An additional LED output type has been added that is called SPI LED co processor. This is a test of outputting LED data to a different device other than what's built-in, a special SPI LED co processor receiving pixel data over SPI and then out putting it in a modular way. This is the first implementation, not complete and functioning properly yet, but we're getting there.